norice

+++ back/Giboo/src/main/java/kh/fin/giboo/alarm/model/service/AlarmService.java
@@ -16,7 +16,7 @@ public interface AlarmService {
 	Map<String, Object> selectAll(Model model);
 	
 
-	int updateReadStatus(double alarmNo, String readStatus);
+	int updateReadStatus(int alarmNo, String readStatus);
